requirements.txt
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.36.0
pyyaml>=6.0.1
scikit-learn>=1.3.0
plotly>=5.22.0
openai>=1.51.0
python-dotenv>=1.0.1
